make "config" for Hardware serial a uint32_t type

+++ cores/stm32l4/USBAPI.h
@@ -59,7 +59,7 @@ class CDC : public HardwareSerial
 public:
     CDC(struct _stm32l4_usbd_cdc_t *usbd_cdc, bool serialEvent);
     void begin(unsigned long baudRate);
-    void begin(unsigned long baudrate, uint16_t config);
+    void begin(unsigned long baudrate, uint32_t config);
     void end(void);
 
     int available(void);
